eslint bug fix
