+++ frontend/src/components/context/AuthContext.tsx
@@ -1,11 +1,5 @@
 // components/context/AuthContext.tsx
-import {
-  createContext,
-  useContext,
-  useEffect,
-  useState,
-  useCallback,
-} from "react";
+import { createContext } from "react";
 
 interface AuthContextType {
   isLoggedIn: boolean | null;
@@ -13,35 +7,8 @@ interface AuthContextType {
   refresh: () => Promise<void>;
 }
 
-const AuthContext = createContext<AuthContextType>({
+export const AuthContext = createContext<AuthContextType>({
   isLoggedIn: null,
   loading: true,
   refresh: async () => {},
 });
-
-export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
-  const [isLoggedIn, setIsLoggedIn] = useState<boolean | null>(null);
-
-  const refresh = useCallback(async () => {
-    try {
-      const res = await fetch("/api/profile", { credentials: "include" });
-      setIsLoggedIn(res.ok);
-    } catch {
-      setIsLoggedIn(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    refresh();
-  }, [refresh]);
-
-  return (
-    <AuthContext.Provider
-      value={{ isLoggedIn, loading: isLoggedIn === null, refresh }}
-    >
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => useContext(AuthContext);
