+++ frontend/src/components/context/AuthProvider.tsx
@@ -1,0 +1,27 @@
+import { useCallback, useEffect, useState } from "react";
+import { AuthContext } from "./AuthContext";
+
+export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
+  const [isLoggedIn, setIsLoggedIn] = useState<boolean | null>(null);
+
+  const refresh = useCallback(async () => {
+    try {
+      const res = await fetch("/api/profile", { credentials: "include" });
+      setIsLoggedIn(res.ok);
+    } catch {
+      setIsLoggedIn(false);
+    }
+  }, []);
+
+  useEffect(() => {
+    refresh();
+  }, [refresh]);
+
+  return (
+    <AuthContext.Provider
+      value={{ isLoggedIn, loading: isLoggedIn === null, refresh }}
+    >
+      {children}
+    </AuthContext.Provider>
+  );
+};
