+++ frontend/src/components/pages/LoginPage.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import "./../../index.css";
 import { loginUser } from "../../authentication/login";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/useAuth";
 
 const LoginPage: React.FC = () => {
   const [email, setEmail] = useState("");
