+++ frontend/src/components/wrappers/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "./../context/AuthContext";
+import { useAuth } from "./../context/useAuth";
 import { JSX } from "react";
 
 interface ProtectedRouteProps {
